Parse message bodies with complex types (#4)
We assumed initially that every field in an SNS message would be a
string. However some of them can be complex objects. We can avoid
having to figure that out by just parsing them all to json.RawMessage
and then parsing the only two fields we actually need to strings.